pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bedrock-latency-benchmark"
version = "0.1.0"
description = "AWS Bedrock lightweight model latency benchmark tool"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "boto3>=1.34.0",
    "strands-agents>=0.1.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "hypothesis>=6.0.0",
]

[project.scripts]
bedrock-benchmark = "src.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
